requirements.txt
numpy==1.18.0
pandas==0.24.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
